clear up OService user flag and status constants
In 34e7946 we added a bunch of new user statuses. We observed that there
were some duplicate values in the list. Turns out the "duplicates" were
flags incorrectly named as statuses.

Digging further into the protocol docs, it appears there are two
different kinds of constants: flags and statuses. Flags are uint16,
statuses are uin32.

Cleaned up some errors and inconsistencies:
- Status consts were incorrectly namd as flags.
- Two flags were typed as uint32 when they should hace remained uint16.
- Some of the status const values are not powers of 2, not sure how they
  can be useful as flags.
- User info TLV cast the invisibility flag, a uint32, to uint16. The
  invisibility flagged still worked despite this. Re-typed the flag to
  uint32 for consistency's sake.

+++ state/session_test.go
@@ -83,7 +83,7 @@ func TestSession_TLVUserInfo(t *testing.T) {
 					TLVList: wire.TLVList{
 						wire.NewTLV(wire.OServiceUserInfoSignonTOD, uint32(1)),
 						wire.NewTLV(wire.OServiceUserInfoUserFlags, uint16(0x0010)),
-						wire.NewTLV(wire.OServiceUserInfoStatus, uint16(0x0000)),
+						wire.NewTLV(wire.OServiceUserInfoStatus, uint32(0x0000)),
 						wire.NewTLV(wire.OServiceUserInfoIdleTime, uint16(0)),
 					},
 				},
@@ -102,7 +102,7 @@ func TestSession_TLVUserInfo(t *testing.T) {
 					TLVList: wire.TLVList{
 						wire.NewTLV(wire.OServiceUserInfoSignonTOD, uint32(1)),
 						wire.NewTLV(wire.OServiceUserInfoUserFlags, uint16(0x30)),
-						wire.NewTLV(wire.OServiceUserInfoStatus, uint16(0x0000)),
+						wire.NewTLV(wire.OServiceUserInfoStatus, uint32(0x0000)),
 						wire.NewTLV(wire.OServiceUserInfoIdleTime, uint16(0)),
 					},
 				},
@@ -121,7 +121,7 @@ func TestSession_TLVUserInfo(t *testing.T) {
 					TLVList: wire.TLVList{
 						wire.NewTLV(wire.OServiceUserInfoSignonTOD, uint32(1)),
 						wire.NewTLV(wire.OServiceUserInfoUserFlags, uint16(0x0010)),
-						wire.NewTLV(wire.OServiceUserInfoStatus, uint16(0x0100)),
+						wire.NewTLV(wire.OServiceUserInfoStatus, uint32(0x0100)),
 						wire.NewTLV(wire.OServiceUserInfoIdleTime, uint16(0)),
 					},
 				},
@@ -144,7 +144,7 @@ func TestSession_TLVUserInfo(t *testing.T) {
 					TLVList: wire.TLVList{
 						wire.NewTLV(wire.OServiceUserInfoSignonTOD, uint32(1)),
 						wire.NewTLV(wire.OServiceUserInfoUserFlags, uint16(0x0010)),
-						wire.NewTLV(wire.OServiceUserInfoStatus, uint16(0x0000)),
+						wire.NewTLV(wire.OServiceUserInfoStatus, uint32(0x0000)),
 						wire.NewTLV(wire.OServiceUserInfoIdleTime, uint16(1001)),
 					},
 				},
@@ -164,7 +164,7 @@ func TestSession_TLVUserInfo(t *testing.T) {
 					TLVList: wire.TLVList{
 						wire.NewTLV(wire.OServiceUserInfoSignonTOD, uint32(1)),
 						wire.NewTLV(wire.OServiceUserInfoUserFlags, uint16(0x0010)),
-						wire.NewTLV(wire.OServiceUserInfoStatus, uint16(0x0000)),
+						wire.NewTLV(wire.OServiceUserInfoStatus, uint32(0x0000)),
 						wire.NewTLV(wire.OServiceUserInfoIdleTime, uint16(0)),
 					},
 				},
@@ -194,7 +194,7 @@ func TestSession_TLVUserInfo(t *testing.T) {
 					TLVList: wire.TLVList{
 						wire.NewTLV(wire.OServiceUserInfoSignonTOD, uint32(1)),
 						wire.NewTLV(wire.OServiceUserInfoUserFlags, uint16(0x0010)),
-						wire.NewTLV(wire.OServiceUserInfoStatus, uint16(0x0000)),
+						wire.NewTLV(wire.OServiceUserInfoStatus, uint32(0x0000)),
 						wire.NewTLV(wire.OServiceUserInfoIdleTime, uint16(0)),
 						wire.NewTLV(wire.OServiceUserInfoOscarCaps, []byte{
 							// chat: "748F2420-6287-11D1-8222-444553540000"
@@ -221,7 +221,7 @@ func TestSession_TLVUserInfo(t *testing.T) {
 					TLVList: wire.TLVList{
 						wire.NewTLV(wire.OServiceUserInfoSignonTOD, uint32(1)),
 						wire.NewTLV(wire.OServiceUserInfoUserFlags, uint16(0x0010)),
-						wire.NewTLV(wire.OServiceUserInfoStatus, uint16(0x0000)),
+						wire.NewTLV(wire.OServiceUserInfoStatus, uint32(0x0000)),
 						wire.NewTLV(wire.OServiceUserInfoIdleTime, uint16(0)),
 					},
 				},
